# Improve precision of abstract_iteration