@Keno 2020-07-03
> Also note that while this is a nice improvement, performance is still
> not great, since we currently can't inline these apply calls. However,
> fixing that is in progress in a parallel PR and with both changes
> put together, the performance improvement of splatting of small
> iterators is quite sizeable.

Some numbers on this:

Before
```
julia> f(x) = (x...,)
f (generic function with 1 method)

julia> @benchmark f(1=>2)
BenchmarkTools.Trial: 
  memory estimate:  96 bytes
  allocs estimate:  3
  --------------
  minimum time:     351.580 ns (0.00% GC)
  median time:      357.165 ns (0.00% GC)
  mean time:        373.367 ns (0.97% GC)
  maximum time:     9.967 μs (96.17% GC)
  --------------
  samples:          10000
  evals/sample:     212
```
After:
```
julia> @benchmark f(1=>2)
BenchmarkTools.Trial: 
  memory estimate:  0 bytes
  allocs estimate:  0
  --------------
  minimum time:     2.026 ns (0.00% GC)
  median time:      2.032 ns (0.00% GC)
  mean time:        2.035 ns (0.00% GC)
  maximum time:     6.816 ns (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     1000
```