@martinholters 2020-07-03
Something I long had on my to do list, but a half-hearted attempt didn't work out, so it stayed there. Glad you tackled it!

I wouldn't worry too much about the high MAX_TUPLE_SPLAT: For those cases where this PR makes a difference, i.e. where `iterate` returns a const-inferable state, it is also probably quite simple and therefore cheap to infer/const-propagate. Depending on the context, the more precise inference result might make up for the extra work. The worst case is probably where MAX_TUPLE_SPLAT is exceeded: Extra work with little profit. Code splatting lots of large StaticArrays might take a hit? A middle ground might be to have a second, lower cut-off and start `widenconst`ing (i.e. like before this PR) once that is exceeded. But let's first see whether this causes any relevant latency issues as is.