@martinholters 2020-07-03
I've noticed something else funny going on here: After `include("compiler/compiler.jl")`, displaying the result of splatting into tuple produces
```
Internal error: encountered unexpected error in runtime:
TypeError(func=:Tuple, context="parameter", expected=Type, got=Core.Compiler.Const(val=2, actual=false))
jl_type_error_rt at ../src/rtutils.c:118
jl_f_apply_type at ../src/builtins.c:962
jl_apply at ../src/julia.h:1712 [inlined]
do_apply at ../src/builtins.c:655
argtypes_to_type at ./compiler/typeutils.jl:48
abstract_call_known at ./compiler/abstractinterpretation.jl:931
abstract_call at ./compiler/abstractinterpretation.jl:955
abstract_call at ./compiler/abstractinterpretation.jl:939
abstract_eval_statement at ./compiler/abstractinterpretation.jl:1061
typeinf_local at ./compiler/abstractinterpretation.jl:1300
typeinf_nocycle at ./compiler/abstractinterpretation.jl:1371
typeinf at ./compiler/typeinfer.jl:12
abstract_call_method_with_const_args at ./compiler/abstractinterpretation.jl:289
unknown function (ip: 0x7f6dcb9b1af7)
abstract_call_gf_by_type at ./compiler/abstractinterpretation.jl:139
abstract_call_known at ./compiler/abstractinterpretation.jl:932
abstract_call at ./compiler/abstractinterpretation.jl:955
[...]
jl_compile_method_internal at ../src/gf.c:2074
jl_compile_method_internal at ../src/gf.c:2037 [inlined]
_jl_invoke at ../src/gf.c:2344 [inlined]
jl_apply_generic at ../src/gf.c:2535
display at ../stdlib/v1.6/REPL/src/REPL.jl:218
display at ../stdlib/v1.6/REPL/src/REPL.jl:232
display at ./multimedia.jl:328
```
E.g.
```julia
julia> f() = (()...,)
f (generic function with 1 method)

julia> f();

julia> f()
Internal error: encountered unexpected error in runtime:
TypeError(func=:Tuple, context="parameter", expected=Type, got=Core.Compiler.Const(val=2, actual=false))
[...]
()

julia> @code_typed f() # also here
CodeInfo(
1 ─     return Internal error: encountered unexpected error in runtime:
TypeError(func=:Tuple, context="parameter", expected=Type, got=Core.Compiler.Const(val=2, actual=false))
[...]
()
) => Tuple{}

```

The `argtypes` given to `argtypes_to_type` in these cases are
```julia
Array{Any, (3,)}[
  Core.Compiler.Const(val=typeof(Base.iterate)(), actual=false),
  Core.Compiler.Const(val=("#= circular reference @-", 1, " =#"), actual=false),
  Core.Compiler.Const(val=2, actual=false)]
```
and get widened to
```julia
Array{Any, (3,)}[
  typeof(Base.iterate),
  Tuple{String, Int64, String},
  Core.Compiler.Const(val=2, actual=false)]
```
No idea what to make of this...