@martinholters 2020-07-03
Ah, maybe the `Core.Compiler.Const(val=2, actual=false)` is a result cached from before reloading `compiler.jl` and hence is not recognized by `widenconst`? And this is to be expected and `include("compiler/compiler.jl")` is just a don't-do-that? If so sorry for the noise.