@Keno 2020-07-03
> After `include("compiler/compiler.jl")`, displaying the result of splatting into tuple produces

I can't reproduce this. I usually just use `using Revise; Revise.track(Core.Compiler)` which works fine most of the time.