@Keno 2020-07-03
Anyway, I rotated the second loop, which I think should address the bug you found.