@Keno 2020-07-08
> I've noticed something else funny going on here: After `include("compiler/compiler.jl")`, displaying the result of splatting into tuple produces

This will be fixed by #36570